infiltration_clients/requirements.txt
backcall==0.1.0
certifi==2019.3.9
chardet==3.0.4
decorator==4.4.0
idna==2.8
ipython==7.4.0
ipython-genutils==0.2.0
jedi==0.13.3
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
pika==1.0.1
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pycrypto==2.6.1
Pygments==2.3.1
PyYAML==5.1
requests==2.21.0
six==1.12.0
traitlets==4.3.2
urllib3==1.24.2
wcwidth==0.1.7
